refactor: rename for event handler directory
