+++ src/input_device_monitor/mod.rs
@@ -1,4 +1,6 @@
 pub mod sender;
-pub mod event;
+pub mod event_handler;
+pub mod event_caster;
+pub mod my_event;
 pub mod app;
